requirements.txt
python-dotenv==1.1.1
streamlit==1.49.1
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
openai==1.107.1
